bypass ad-hoc solution with killing all nginx processes in ubuntu 16.04

+++ recipes/default.rb
@@ -28,6 +28,7 @@ begin
   Chef::Log.debug("Resources found inside recipe[#{cookbook_name}::#{recipe_name}]: #{this_recipe_resource_collection.length}")
   fail Chef::Exceptions::ResourceNotFound if this_recipe_resource_collection.length == 0
 rescue Chef::Exceptions::ResourceNotFound
+
   execute 'kill nginx after installation' do
     command '/usr/bin/pkill nginx'
     action :nothing
@@ -35,7 +36,9 @@ rescue Chef::Exceptions::ResourceNotFound
   end
 
   package 'nginx' do
-    notifies :run, 'execute[kill nginx after installation]'
+    if node['platform_version'].to_f < 16.04
+      notifies :run, 'execute[kill nginx after installation]'
+    end
   end
 
   directory node['nginx']['directories']['log_dir'] do
